Refactor: [BADA-105] 파일 위치 변경 (trade/detail/..) #46
