+++ src/features/trade/components/detail/TradeDetailDrawer.tsx
