+++ src/features/trade/components/detail/TradeDetailProductSection.tsx
