+++ src/features/trade/components/detail/TradeDetailSellerSection.tsx
